pyproject.toml
[project]
name = "tbc-app-services-cron"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
